+++ back-end/Changelog.md
@@ -8,3 +8,4 @@
 - Create author,handle validation and exception handling
 - Create get all authors function and implement dto
 - create get author by id function
+- make change the find author by id function